cicle membership work

+++ Youverse.Hosting/Controllers/OwnerToken/Circles/CircleDefinitionController.cs
@@ -3,7 +3,9 @@ using System.Collections.Generic;
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.Mvc;
 using Youverse.Core;
+using Youverse.Core.Exceptions;
 using Youverse.Core.Services.Contacts.Circle.Definition;
+using Youverse.Core.Services.Contacts.Circle.Membership;
 
 namespace Youverse.Hosting.Controllers.OwnerToken.Circles
 {
@@ -13,10 +15,12 @@ namespace Youverse.Hosting.Controllers.OwnerToken.Circles
     public class CircleDefinitionController : ControllerBase
     {
         private readonly CircleDefinitionService _circleDefinitionService;
+        private readonly ICircleNetworkService _cns;
 
-        public CircleDefinitionController(CircleDefinitionService circleDefinitionService)
+        public CircleDefinitionController(CircleDefinitionService circleDefinitionService, ICircleNetworkService cns)
         {
             _circleDefinitionService = circleDefinitionService;
+            _cns = cns;
         }
 
         [HttpGet("list")]
@@ -49,6 +53,13 @@ namespace Youverse.Hosting.Controllers.OwnerToken.Circles
         [HttpPost("delete")]
         public async Task<bool> DeleteCircle([FromBody] Guid id)
         {
+            //TODO: not too much a fan of this being in a controller but breaking out a whole other class for this requires more need 
+            var canDelete = await _cns.CanDeleteCircle(new ByteArrayId(id.ToByteArray()));
+            if (!canDelete)
+            {
+                throw new YouverseException("Cannot delete a circle with members");
+            }
+            
             await _circleDefinitionService.Delete(id);
             return true;
         }
